load data to frontend components

+++ internal/core/components/position.go
@@ -2,12 +2,12 @@ package components
 
 // Position represents a champion's position on the board
 type Position struct {
-	x float64
-	y float64
+	x int
+	y int
 }
 
 // NewPosition creates a Position component
-func NewPosition(x, y float64) Position {
+func NewPosition(x, y int) Position {
 	return Position{
 		x: x,
 		y: y,
@@ -15,27 +15,27 @@ func NewPosition(x, y float64) Position {
 }
 
 // GetX returns the x-coordinate of the position
-func (p Position) GetX() float64 {
+func (p Position) GetX() int {
 	return p.x
 }
 
 // GetY returns the y-coordinate of the position
-func (p Position) GetY() float64 {
+func (p Position) GetY() int {
 	return p.y
 }
 
 // SetX sets the x-coordinate of the position
-func (p *Position) SetX(x float64) {
+func (p *Position) SetX(x int) {
 	p.x = x
 }
 
 // SetY sets the y-coordinate of the position
-func (p *Position) SetY(y float64) {
+func (p *Position) SetY(y int) {
 	p.y = y
 }
 
 // SetPosition sets both x and y coordinates of the position
-func (p *Position) SetPosition(x, y float64) {
+func (p *Position) SetPosition(x, y int) {
 	p.x = x
 	p.y = y
 }
